pyproject.toml
[tool.poetry]
name = "tdd project"
version = "0.0.1"
description = ""
authors = ["KhovetS2 <isaque.elis.escola@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.11"}
pydantic-settings = "^2.8.1"
pytest = "^8.3.5"
motor = "^3.7.0"
pytest-asyncio = "^0.25.3"
pre-commit = "^4.1.0"
httpx = "^0.28.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = [
  "--strict-config",
  "--strict-markers",
  "--ignore=docs_src",
]
xfail_strict = true
junit_family = "xunit2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
